added tv list state
